+++ issuing_cardholder.go
@@ -38,17 +38,55 @@ type IssuingBillingParams struct {
 	Name    *string        `form:"name"`
 }
 
+// IssuingCardholderCompanyParams represents additional information about a
+// `business_entity` cardholder.
+type IssuingCardholderCompanyParams struct {
+	TaxID *string `form:"tax_id"`
+}
+
+// IssuingCardholderIndividualVerificationDocumentParams represents an
+// identifying document, either a passport or local ID card.
+type IssuingCardholderIndividualVerificationDocumentParams struct {
+	Back  *string `form:"back"`
+	Front *string `form:"front"`
+}
+
+// IssuingCardholderIndividualVerificationParams represents government-issued ID
+// document for this cardholder.
+type IssuingCardholderIndividualVerificationParams struct {
+	Document *IssuingCardholderIndividualVerificationDocumentParams `form:"document"`
+}
+
+// IssuingCardholderIndividualDOBParams represents the date of birth of the
+// cardholder individual.
+type IssuingCardholderIndividualDOBParams struct {
+	Day   *int64 `form:"day"`
+	Month *int64 `form:"month"`
+	Year  *int64 `form:"year"`
+}
+
+// IssuingCardholderIndividualParams represents additional information about an
+// `individual` cardholder.
+type IssuingCardholderIndividualParams struct {
+	DOB          *IssuingCardholderIndividualDOBParams          `form:"dob"`
+	FirstName    *string                                        `form:"first_name"`
+	LastName     *string                                        `form:"last_name"`
+	Verification *IssuingCardholderIndividualVerificationParams `form:"verification"`
+}
+
 // IssuingCardholderParams is the set of parameters that can be used when creating or updating an issuing cardholder.
 type IssuingCardholderParams struct {
 	Params                `form:"*"`
-	AuthorizationControls *AuthorizationControlsParams `form:"authorization_controls"`
-	Billing               *IssuingBillingParams        `form:"billing"`
-	Email                 *string                      `form:"email"`
-	IsDefault             *bool                        `form:"is_default"`
-	Name                  *string                      `form:"name"`
-	PhoneNumber           *string                      `form:"phone_number"`
-	Status                *string                      `form:"status"`
-	Type                  *string                      `form:"type"`
+	AuthorizationControls *AuthorizationControlsParams       `form:"authorization_controls"`
+	Billing               *IssuingBillingParams              `form:"billing"`
+	Company               *IssuingCardholderCompanyParams    `form:"company"`
+	Email                 *string                            `form:"email"`
+	Individual            *IssuingCardholderIndividualParams `form:"individual"`
+	IsDefault             *bool                              `form:"is_default"`
+	Name                  *string                            `form:"name"`
+	PhoneNumber           *string                            `form:"phone_number"`
+	Status                *string                            `form:"status"`
+	Type                  *string                            `form:"type"`
 }
 
 // IssuingCardholderListParams is the set of parameters that can be used when listing issuing cardholders.
@@ -75,13 +113,51 @@ type IssuingCardholderRequirements struct {
 	PastDue        []string                                    `json:"past_due"`
 }
 
+// IssuingCardholderIndividualVerificationDocument represents an identifying
+// document, either a passport or local ID card.
+type IssuingCardholderIndividualVerificationDocument struct {
+	Back  *File `json:"back"`
+	Front *File `json:"front"`
+}
+
+// IssuingCardholderIndividualVerification represents the Government-issued ID
+// document for this cardholder
+type IssuingCardholderIndividualVerification struct {
+	Document *IssuingCardholderIndividualVerificationDocument `json:"document"`
+}
+
+// IssuingCardholderIndividualDOB represents the date of birth of the issuing card hoder
+// individual.
+type IssuingCardholderIndividualDOB struct {
+	Day   int64 `json:"day"`
+	Month int64 `json:"month"`
+	Year  int64 `json:"year"`
+}
+
+// IssuingCardholderIndividual represents additional information about an
+// individual cardholder.
+type IssuingCardholderIndividual struct {
+	DOB          *IssuingCardholderIndividualDOB          `json:"dob"`
+	FirstName    string                                   `json:"first_name"`
+	LastName     string                                   `json:"last_name"`
+	Verification *IssuingCardholderIndividualVerification `json:"verification"`
+}
+
+// IssuingCardholderCompany represents additional information about a
+// business_entity cardholder.
+type IssuingCardholderCompany struct {
+	TaxIDProvided bool `json:"tax_id_provided"`
+}
+
 // IssuingCardholder is the resource representing a Stripe issuing cardholder.
 type IssuingCardholder struct {
 	AuthorizationControls *IssuingCardAuthorizationControls `json:"authorization_controls"`
 	Billing               *IssuingBilling                   `json:"billing"`
+	Company               *IssuingCardholderCompany         `json:"company"`
 	Created               int64                             `json:"created"`
 	Email                 string                            `json:"email"`
 	ID                    string                            `json:"id"`
+	Individual            *IssuingCardholderIndividual      `json:"individual"`
 	Livemode              bool                              `json:"livemode"`
 	Metadata              map[string]string                 `json:"metadata"`
 	Name                  string                            `json:"name"`